Moved time-related utility functions to TimeScale
