+++ src/main/java/org/radarcns/mongo/data/passive/SourceDataMongoWrapper.java
@@ -38,6 +38,7 @@ import org.radarcns.domain.restapi.header.TimeFrame;
 import org.radarcns.mongo.util.MongoHelper;
 import org.radarcns.mongo.util.MongoHelper.Stat;
 import org.radarcns.util.RadarConverter;
+import org.radarcns.util.TimeScale;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -242,10 +243,9 @@ public abstract class SourceDataMongoWrapper {
      */
     protected abstract int extractCount(Document doc);
 
-
     public Double getExpectedRecordCount(TimeWindow timeWindow) {
-        return RadarConverter.getSecond(timeWindow) * Double
-                .valueOf(this.sourceData.getFrequency());
+        return RadarConverter.getExpectedMessages(timeWindow,
+                Double.valueOf(this.sourceData.getFrequency());
     }
 
     /**
